CSI-3103 move main from controller server (#355)
move main from controller server
