+++ controller/scripts/entrypoint.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-exec python3.8 /driver/controller/controller_server/csi_controller_server.py $@
+exec python3.8 /driver/controller/controller_server/main.py $@
