+++ controller/tests/controller_server/utils_test.py
@@ -17,8 +17,7 @@ from controller.tests.utils import get_fake_secret_config
 class TestUtils(unittest.TestCase):
 
     def setUp(self):
-        self.fqdn = "fqdn"
-        self.servicer = ControllerServicer(self.fqdn)
+        self.servicer = ControllerServicer()
         self.config = {"controller": {"publish_context_lun_parameter": "lun",
                                       "publish_context_connectivity_parameter": "connectivity_type",
                                       "publish_context_separator": ",",
